Clean up of list sorting.

+++ include/lgi/common/List.h
@@ -271,7 +271,7 @@ protected:
 	LRect &GetClientRect();
 	void PourAll();
 	void UpdateScrollBars();
-	void KeyScroll(int iTo, int iFrom, bool SelectItems);
+	void KeyScroll(ssize_t iTo, ssize_t iFrom, bool SelectItems);
 	void ClearDs(int Col) override;
 	ContainerItemDrop GetItemReorderPos(LPoint ms) override;
 
@@ -466,7 +466,7 @@ public:
 		if (!Compare || !Lock(_FL))
 			return;
 
-		LListItem *Kb = Keyboard >= 0 && Keyboard < Items.Length() ? Items[Keyboard] : NULL;
+		LListItem *Kb = Keyboard >= 0 && Keyboard < (ssize_t)Items.Length() ? Items[Keyboard] : NULL;
 		Items.Sort([Compare, Data](auto a, auto b)
 		{
 			return Compare(a, b, Data);
@@ -486,7 +486,7 @@ public:
 		if (!compare || !Lock(_FL))
 			return;
 
-		LListItem *Kb = Keyboard >= 0 && Keyboard < Items.Length() ? Items[Keyboard] : NULL;
+		LListItem *Kb = Keyboard >= 0 && Keyboard < (ssize_t)Items.Length() ? Items[Keyboard] : NULL;
 		Items.Sort(compare);
 		Keyboard = Kb ? Items.IndexOf(Kb) : -1;
 		Unlock();
